+++ rosariosis.sql
@@ -3499,7 +3499,7 @@ INSERT INTO address VALUES (0, NULL, NULL, NULL, NULL, NULL, NULL, NULL, NULL, N
 -- Data for Name: attendance_calendars; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO attendance_calendars VALUES (1, 'Main', 2017, 1, 'Y', NULL);
+INSERT INTO attendance_calendars VALUES (1, 'Main', 2018, 1, 'Y', NULL);
 
 
 --
@@ -3512,10 +3512,10 @@ INSERT INTO attendance_calendars VALUES (1, 'Main', 2017, 1, 'Y', NULL);
 -- Data for Name: attendance_codes; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO attendance_codes VALUES (1, 2017, 1, 'Absent', 'A', 'teacher', 'A', NULL, 0, NULL);
-INSERT INTO attendance_codes VALUES (2, 2017, 1, 'Present', 'P', 'teacher', 'P', 'Y', 0, NULL);
-INSERT INTO attendance_codes VALUES (3, 2017, 1, 'Tardy', 'T', 'teacher', 'P', NULL, 0, NULL);
-INSERT INTO attendance_codes VALUES (4, 2017, 1, 'Excused Absence', 'E', 'official', 'A', NULL, 0, NULL);
+INSERT INTO attendance_codes VALUES (1, 2018, 1, 'Absent', 'A', 'teacher', 'A', NULL, 0, NULL);
+INSERT INTO attendance_codes VALUES (2, 2018, 1, 'Present', 'P', 'teacher', 'P', 'Y', 0, NULL);
+INSERT INTO attendance_codes VALUES (3, 2018, 1, 'Tardy', 'T', 'teacher', 'P', NULL, 0, NULL);
+INSERT INTO attendance_codes VALUES (4, 2018, 1, 'Excused Absence', 'E', 'official', 'A', NULL, 0, NULL);
 
 
 --
@@ -3644,10 +3644,10 @@ INSERT INTO custom_fields VALUES (200000011, 'textarea', 'Doctor''s Note Comment
 -- Data for Name: discipline_field_usage; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO discipline_field_usage VALUES (1, 3, 2017, 1, 'Parents Contacted By Teacher', '', 4);
-INSERT INTO discipline_field_usage VALUES (2, 4, 2017, 1, 'Parent Contacted by Administrator', '', 5);
-INSERT INTO discipline_field_usage VALUES (3, 6, 2017, 1, 'Comments', '', 6);
-INSERT INTO discipline_field_usage VALUES (4, 1, 2017, 1, 'Violation', 'Skipping Class
+INSERT INTO discipline_field_usage VALUES (1, 3, 2018, 1, 'Parents Contacted By Teacher', '', 4);
+INSERT INTO discipline_field_usage VALUES (2, 4, 2018, 1, 'Parent Contacted by Administrator', '', 5);
+INSERT INTO discipline_field_usage VALUES (3, 6, 2018, 1, 'Comments', '', 6);
+INSERT INTO discipline_field_usage VALUES (4, 1, 2018, 1, 'Violation', 'Skipping Class
 Profanity, vulgarity, offensive language
 Insubordination (Refusal to Comply, Disrespectful Behavior)
 Inebriated (Alcohol or Drugs)
@@ -3656,11 +3656,11 @@ Harassment
 Fighting
 Public Display of Affection
 Other', 1);
-INSERT INTO discipline_field_usage VALUES (5, 2, 2017, 1, 'Detention Assigned', '10 Minutes
+INSERT INTO discipline_field_usage VALUES (5, 2, 2018, 1, 'Detention Assigned', '10 Minutes
 20 Minutes
 30 Minutes
 Discuss Suspension', 2);
-INSERT INTO discipline_field_usage VALUES (6, 5, 2017, 1, 'Suspensions (Office Only)', 'Half Day
+INSERT INTO discipline_field_usage VALUES (6, 5, 2018, 1, 'Suspensions (Office Only)', 'Half Day
 In School Suspension
 1 Day
 2 Days
@@ -3698,9 +3698,9 @@ INSERT INTO discipline_fields VALUES (6, 'Comments', '', 'textarea', 'CATEGORY_6
 -- Data for Name: eligibility_activities; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO eligibility_activities VALUES (1, 2017, 1, 'Boy''s Basketball', '2017-10-02', '2018-04-13');
-INSERT INTO eligibility_activities VALUES (2, 2017, 1, 'Chess Team', '2017-09-01', '2018-06-08');
-INSERT INTO eligibility_activities VALUES (3, 2017, 1, 'Girl''s Basketball', '2017-10-02', '2018-04-13');
+INSERT INTO eligibility_activities VALUES (1, 2018, 1, 'Boy''s Basketball', '2018-10-01', '2019-04-12');
+INSERT INTO eligibility_activities VALUES (2, 2018, 1, 'Chess Team', '2018-09-03', '2019-06-07');
+INSERT INTO eligibility_activities VALUES (3, 2018, 1, 'Girl''s Basketball', '2018-10-01', '2019-04-12');
 
 
 --
@@ -4141,31 +4141,31 @@ INSERT INTO profile_exceptions VALUES (1, 'Students/StudentBreakdown.php', 'Y', 
 -- Data for Name: program_config; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO program_config VALUES (2017, 1, 'eligibility', 'START_DAY', '1');
-INSERT INTO program_config VALUES (2017, 1, 'eligibility', 'START_HOUR', '23');
-INSERT INTO program_config VALUES (2017, 1, 'eligibility', 'START_MINUTE', '30');
-INSERT INTO program_config VALUES (2017, 1, 'eligibility', 'START_M', 'PM');
-INSERT INTO program_config VALUES (2017, 1, 'eligibility', 'END_DAY', '5');
-INSERT INTO program_config VALUES (2017, 1, 'eligibility', 'END_HOUR', '23');
-INSERT INTO program_config VALUES (2017, 1, 'eligibility', 'END_MINUTE', '30');
-INSERT INTO program_config VALUES (2017, 1, 'eligibility', 'END_M', 'PM');
-INSERT INTO program_config VALUES (2017, 1, 'attendance', 'ATTENDANCE_EDIT_DAYS_BEFORE', NULL);
-INSERT INTO program_config VALUES (2017, 1, 'attendance', 'ATTENDANCE_EDIT_DAYS_AFTER', NULL);
-INSERT INTO program_config VALUES (2017, 1, 'grades', 'GRADES_DOES_LETTER_PERCENT', '0');
-INSERT INTO program_config VALUES (2017, 1, 'grades', 'GRADES_HIDE_NON_ATTENDANCE_COMMENT', NULL);
-INSERT INTO program_config VALUES (2017, 1, 'grades', 'GRADES_TEACHER_ALLOW_EDIT', NULL);
-INSERT INTO program_config VALUES (2017, 1, 'grades', 'GRADES_DO_STATS_STUDENTS_PARENTS', NULL);
-INSERT INTO program_config VALUES (2017, 1, 'grades', 'GRADES_DO_STATS_ADMIN_TEACHERS', 'Y');
-INSERT INTO program_config VALUES (2017, 1, 'students', 'STUDENTS_USE_BUS', 'Y');
-INSERT INTO program_config VALUES (2017, 1, 'students', 'STUDENTS_USE_CONTACT', 'Y');
-INSERT INTO program_config VALUES (2017, 1, 'students', 'STUDENTS_SEMESTER_COMMENTS', NULL);
-INSERT INTO program_config VALUES (2017, 1, 'moodle', 'MOODLE_URL', NULL);
-INSERT INTO program_config VALUES (2017, 1, 'moodle', 'MOODLE_TOKEN', NULL);
-INSERT INTO program_config VALUES (2017, 1, 'moodle', 'MOODLE_PARENT_ROLE_ID', NULL);
-INSERT INTO program_config VALUES (2017, 1, 'moodle', 'ROSARIO_STUDENTS_EMAIL_FIELD_ID', NULL);
-INSERT INTO program_config VALUES (2017, 1, 'food_service', 'FOOD_SERVICE_BALANCE_WARNING', '5');
-INSERT INTO program_config VALUES (2017, 1, 'food_service', 'FOOD_SERVICE_BALANCE_MINIMUM', '-40');
-INSERT INTO program_config VALUES (2017, 1, 'food_service', 'FOOD_SERVICE_BALANCE_TARGET', '19');
+INSERT INTO program_config VALUES (2018, 1, 'eligibility', 'START_DAY', '1');
+INSERT INTO program_config VALUES (2018, 1, 'eligibility', 'START_HOUR', '23');
+INSERT INTO program_config VALUES (2018, 1, 'eligibility', 'START_MINUTE', '30');
+INSERT INTO program_config VALUES (2018, 1, 'eligibility', 'START_M', 'PM');
+INSERT INTO program_config VALUES (2018, 1, 'eligibility', 'END_DAY', '5');
+INSERT INTO program_config VALUES (2018, 1, 'eligibility', 'END_HOUR', '23');
+INSERT INTO program_config VALUES (2018, 1, 'eligibility', 'END_MINUTE', '30');
+INSERT INTO program_config VALUES (2018, 1, 'eligibility', 'END_M', 'PM');
+INSERT INTO program_config VALUES (2018, 1, 'attendance', 'ATTENDANCE_EDIT_DAYS_BEFORE', NULL);
+INSERT INTO program_config VALUES (2018, 1, 'attendance', 'ATTENDANCE_EDIT_DAYS_AFTER', NULL);
+INSERT INTO program_config VALUES (2018, 1, 'grades', 'GRADES_DOES_LETTER_PERCENT', '0');
+INSERT INTO program_config VALUES (2018, 1, 'grades', 'GRADES_HIDE_NON_ATTENDANCE_COMMENT', NULL);
+INSERT INTO program_config VALUES (2018, 1, 'grades', 'GRADES_TEACHER_ALLOW_EDIT', NULL);
+INSERT INTO program_config VALUES (2018, 1, 'grades', 'GRADES_DO_STATS_STUDENTS_PARENTS', NULL);
+INSERT INTO program_config VALUES (2018, 1, 'grades', 'GRADES_DO_STATS_ADMIN_TEACHERS', 'Y');
+INSERT INTO program_config VALUES (2018, 1, 'students', 'STUDENTS_USE_BUS', 'Y');
+INSERT INTO program_config VALUES (2018, 1, 'students', 'STUDENTS_USE_CONTACT', 'Y');
+INSERT INTO program_config VALUES (2018, 1, 'students', 'STUDENTS_SEMESTER_COMMENTS', NULL);
+INSERT INTO program_config VALUES (2018, 1, 'moodle', 'MOODLE_URL', NULL);
+INSERT INTO program_config VALUES (2018, 1, 'moodle', 'MOODLE_TOKEN', NULL);
+INSERT INTO program_config VALUES (2018, 1, 'moodle', 'MOODLE_PARENT_ROLE_ID', NULL);
+INSERT INTO program_config VALUES (2018, 1, 'moodle', 'ROSARIO_STUDENTS_EMAIL_FIELD_ID', NULL);
+INSERT INTO program_config VALUES (2018, 1, 'food_service', 'FOOD_SERVICE_BALANCE_WARNING', '5');
+INSERT INTO program_config VALUES (2018, 1, 'food_service', 'FOOD_SERVICE_BALANCE_MINIMUM', '-40');
+INSERT INTO program_config VALUES (2018, 1, 'food_service', 'FOOD_SERVICE_BALANCE_TARGET', '19');
 
 
 --
@@ -4196,37 +4196,37 @@ INSERT INTO program_config VALUES (2017, 1, 'food_service', 'FOOD_SERVICE_BALANC
 -- Data for Name: report_card_comments; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO report_card_comments VALUES (1, 2017, 1, NULL, NULL, NULL, 1, '^n Fails to Meet Course Requirements');
-INSERT INTO report_card_comments VALUES (2, 2017, 1, NULL, NULL, NULL, 2, '^n Comes to ^s Class Unprepared');
-INSERT INTO report_card_comments VALUES (3, 2017, 1, NULL, NULL, NULL, 3, '^n Exerts Positive Influence in Class');
+INSERT INTO report_card_comments VALUES (1, 2018, 1, NULL, NULL, NULL, 1, '^n Fails to Meet Course Requirements');
+INSERT INTO report_card_comments VALUES (2, 2018, 1, NULL, NULL, NULL, 2, '^n Comes to ^s Class Unprepared');
+INSERT INTO report_card_comments VALUES (3, 2018, 1, NULL, NULL, NULL, 3, '^n Exerts Positive Influence in Class');
 
 
 --
 -- Data for Name: report_card_grade_scales; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO report_card_grade_scales VALUES (1, 2017, 1, 'Main', NULL, NULL, NULL, 1, NULL, 4, 0, NULL);
+INSERT INTO report_card_grade_scales VALUES (1, 2018, 1, 'Main', NULL, NULL, NULL, 1, NULL, 4, 0, NULL);
 
 
 --
 -- Data for Name: report_card_grades; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO report_card_grades VALUES (1, 2017, 1, 'A+', 1, 4.00, 97, 'Consistently superior', 1, NULL);
-INSERT INTO report_card_grades VALUES (2, 2017, 1, 'A', 2, 4.00, 93, 'Superior', 1, NULL);
-INSERT INTO report_card_grades VALUES (3, 2017, 1, 'A-', 3, 3.75, 90, NULL, 1, NULL);
-INSERT INTO report_card_grades VALUES (4, 2017, 1, 'B+', 4, 3.50, 87, NULL, 1, NULL);
-INSERT INTO report_card_grades VALUES (5, 2017, 1, 'B', 5, 3.00, 83, 'Above average', 1, NULL);
-INSERT INTO report_card_grades VALUES (6, 2017, 1, 'B-', 6, 2.75, 80, NULL, 1, NULL);
-INSERT INTO report_card_grades VALUES (7, 2017, 1, 'C+', 7, 2.50, 77, NULL, 1, NULL);
-INSERT INTO report_card_grades VALUES (8, 2017, 1, 'C', 8, 2.00, 73, 'Average', 1, NULL);
-INSERT INTO report_card_grades VALUES (9, 2017, 1, 'C-', 9, 1.75, 70, NULL, 1, NULL);
-INSERT INTO report_card_grades VALUES (10, 2017, 1, 'D+', 10, 1.50, 67, NULL, 1, NULL);
-INSERT INTO report_card_grades VALUES (11, 2017, 1, 'D', 11, 1.00, 63, 'Below average', 1, NULL);
-INSERT INTO report_card_grades VALUES (12, 2017, 1, 'D-', 12, 0.75, 60, NULL, 1, NULL);
-INSERT INTO report_card_grades VALUES (13, 2017, 1, 'F', 13, 0.00, 0, 'Failing', 1, NULL);
-INSERT INTO report_card_grades VALUES (14, 2017, 1, 'I', 14, 0.00, 0, 'Incomplete', 1, NULL);
-INSERT INTO report_card_grades VALUES (15, 2017, 1, 'N/A', 15, 0.00, NULL, NULL, 1, NULL);
+INSERT INTO report_card_grades VALUES (1, 2018, 1, 'A+', 1, 4.00, 97, 'Consistently superior', 1, NULL);
+INSERT INTO report_card_grades VALUES (2, 2018, 1, 'A', 2, 4.00, 93, 'Superior', 1, NULL);
+INSERT INTO report_card_grades VALUES (3, 2018, 1, 'A-', 3, 3.75, 90, NULL, 1, NULL);
+INSERT INTO report_card_grades VALUES (4, 2018, 1, 'B+', 4, 3.50, 87, NULL, 1, NULL);
+INSERT INTO report_card_grades VALUES (5, 2018, 1, 'B', 5, 3.00, 83, 'Above average', 1, NULL);
+INSERT INTO report_card_grades VALUES (6, 2018, 1, 'B-', 6, 2.75, 80, NULL, 1, NULL);
+INSERT INTO report_card_grades VALUES (7, 2018, 1, 'C+', 7, 2.50, 77, NULL, 1, NULL);
+INSERT INTO report_card_grades VALUES (8, 2018, 1, 'C', 8, 2.00, 73, 'Average', 1, NULL);
+INSERT INTO report_card_grades VALUES (9, 2018, 1, 'C-', 9, 1.75, 70, NULL, 1, NULL);
+INSERT INTO report_card_grades VALUES (10, 2018, 1, 'D+', 10, 1.50, 67, NULL, 1, NULL);
+INSERT INTO report_card_grades VALUES (11, 2018, 1, 'D', 11, 1.00, 63, 'Below average', 1, NULL);
+INSERT INTO report_card_grades VALUES (12, 2018, 1, 'D-', 12, 0.75, 60, NULL, 1, NULL);
+INSERT INTO report_card_grades VALUES (13, 2018, 1, 'F', 13, 0.00, 0, 'Failing', 1, NULL);
+INSERT INTO report_card_grades VALUES (14, 2018, 1, 'I', 14, 0.00, 0, 'Incomplete', 1, NULL);
+INSERT INTO report_card_grades VALUES (15, 2018, 1, 'N/A', 15, 0.00, NULL, NULL, 1, NULL);
 
 
 --
@@ -4270,46 +4270,46 @@ INSERT INTO school_gradelevels VALUES (9, 1, '08', '8th', NULL, 9);
 -- Data for Name: school_marking_periods; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO school_marking_periods VALUES (1, 2017, 'FY', 1, NULL, 'Full Year', 'FY', 1, '2017-06-12', '2018-06-08', NULL, NULL, NULL, NULL, NULL);
-INSERT INTO school_marking_periods VALUES (2, 2017, 'SEM', 1, 1, 'Semester 1', 'S1', 1, '2017-06-12', '2017-12-31', '2017-12-28', '2017-12-31', NULL, NULL, NULL);
-INSERT INTO school_marking_periods VALUES (3, 2017, 'SEM', 1, 1, 'Semester 2', 'S2', 2, '2018-01-02', '2018-06-08', '2018-06-06', '2018-06-08', NULL, NULL, NULL);
-INSERT INTO school_marking_periods VALUES (4, 2017, 'QTR', 1, 2, 'Quarter 1', 'Q1', 1, '2017-06-12', '2017-09-15', '2017-09-13', '2017-09-15', 'Y', 'Y', NULL);
-INSERT INTO school_marking_periods VALUES (5, 2017, 'QTR', 1, 2, 'Quarter 2', 'Q2', 2, '2017-09-18', '2017-12-31', '2017-12-28', '2017-12-31', 'Y', 'Y', NULL);
-INSERT INTO school_marking_periods VALUES (6, 2017, 'QTR', 1, 3, 'Quarter 3', 'Q3', 3, '2018-01-02', '2018-03-16', '2018-03-14', '2018-03-16', 'Y', 'Y', NULL);
-INSERT INTO school_marking_periods VALUES (7, 2017, 'QTR', 1, 3, 'Quarter 4', 'Q4', 4, '2018-03-19', '2018-06-08', '2018-06-06', '2018-06-08', 'Y', 'Y', NULL);
+INSERT INTO school_marking_periods VALUES (1, 2018, 'FY', 1, NULL, 'Full Year', 'FY', 1, '2018-06-18', '2019-06-07', NULL, NULL, NULL, NULL, NULL);
+INSERT INTO school_marking_periods VALUES (2, 2018, 'SEM', 1, 1, 'Semester 1', 'S1', 1, '2018-06-18', '2018-12-31', '2018-12-28', '2018-12-31', NULL, NULL, NULL);
+INSERT INTO school_marking_periods VALUES (3, 2018, 'SEM', 1, 1, 'Semester 2', 'S2', 2, '2019-01-02', '2019-06-07', '2019-06-05', '2019-06-07', NULL, NULL, NULL);
+INSERT INTO school_marking_periods VALUES (4, 2018, 'QTR', 1, 2, 'Quarter 1', 'Q1', 1, '2018-06-18', '2018-09-14', '2018-09-12', '2018-09-14', 'Y', 'Y', NULL);
+INSERT INTO school_marking_periods VALUES (5, 2018, 'QTR', 1, 2, 'Quarter 2', 'Q2', 2, '2018-09-17', '2018-12-31', '2018-12-28', '2018-12-31', 'Y', 'Y', NULL);
+INSERT INTO school_marking_periods VALUES (6, 2018, 'QTR', 1, 3, 'Quarter 3', 'Q3', 3, '2019-01-02', '2019-03-15', '2019-03-13', '2019-03-15', 'Y', 'Y', NULL);
+INSERT INTO school_marking_periods VALUES (7, 2018, 'QTR', 1, 3, 'Quarter 4', 'Q4', 4, '2019-03-18', '2019-06-07', '2019-06-05', '2019-06-07', 'Y', 'Y', NULL);
 
 
 --
 -- Data for Name: school_periods; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO school_periods VALUES (1, 2017, 1, 1, 'Full Day', 'FD', 300, NULL, NULL, NULL, 'Y', NULL);
-INSERT INTO school_periods VALUES (2, 2017, 1, 2, 'Half Day AM', 'AM', 150, NULL, NULL, NULL, 'Y', NULL);
-INSERT INTO school_periods VALUES (3, 2017, 1, 3, 'Half Day PM', 'PM', 150, NULL, NULL, NULL, 'Y', NULL);
-INSERT INTO school_periods VALUES (4, 2017, 1, 4, 'Period 1', '01', 0, NULL, NULL, NULL, 'Y', NULL);
-INSERT INTO school_periods VALUES (5, 2017, 1, 5, 'Period 2', '02', 0, NULL, NULL, NULL, 'Y', NULL);
-INSERT INTO school_periods VALUES (6, 2017, 1, 6, 'Period 3', '03', 0, NULL, NULL, NULL, 'Y', NULL);
-INSERT INTO school_periods VALUES (7, 2017, 1, 7, 'Period 4', '04', 0, NULL, NULL, NULL, 'Y', NULL);
-INSERT INTO school_periods VALUES (8, 2017, 1, 8, 'Period 5', '05', 0, NULL, NULL, NULL, 'Y', NULL);
-INSERT INTO school_periods VALUES (9, 2017, 1, 9, 'Period 6', '06', 0, NULL, NULL, NULL, 'Y', NULL);
-INSERT INTO school_periods VALUES (10, 2017, 1, 10, 'Period 7', '07', 0, NULL, NULL, NULL, 'Y', NULL);
-INSERT INTO school_periods VALUES (11, 2017, 1, 11, 'Period 8', '08', 0, NULL, NULL, NULL, 'Y', NULL);
+INSERT INTO school_periods VALUES (1, 2018, 1, 1, 'Full Day', 'FD', 300, NULL, NULL, NULL, 'Y', NULL);
+INSERT INTO school_periods VALUES (2, 2018, 1, 2, 'Half Day AM', 'AM', 150, NULL, NULL, NULL, 'Y', NULL);
+INSERT INTO school_periods VALUES (3, 2018, 1, 3, 'Half Day PM', 'PM', 150, NULL, NULL, NULL, 'Y', NULL);
+INSERT INTO school_periods VALUES (4, 2018, 1, 4, 'Period 1', '01', 0, NULL, NULL, NULL, 'Y', NULL);
+INSERT INTO school_periods VALUES (5, 2018, 1, 5, 'Period 2', '02', 0, NULL, NULL, NULL, 'Y', NULL);
+INSERT INTO school_periods VALUES (6, 2018, 1, 6, 'Period 3', '03', 0, NULL, NULL, NULL, 'Y', NULL);
+INSERT INTO school_periods VALUES (7, 2018, 1, 7, 'Period 4', '04', 0, NULL, NULL, NULL, 'Y', NULL);
+INSERT INTO school_periods VALUES (8, 2018, 1, 8, 'Period 5', '05', 0, NULL, NULL, NULL, 'Y', NULL);
+INSERT INTO school_periods VALUES (9, 2018, 1, 9, 'Period 6', '06', 0, NULL, NULL, NULL, 'Y', NULL);
+INSERT INTO school_periods VALUES (10, 2018, 1, 10, 'Period 7', '07', 0, NULL, NULL, NULL, 'Y', NULL);
+INSERT INTO school_periods VALUES (11, 2018, 1, 11, 'Period 8', '08', 0, NULL, NULL, NULL, 'Y', NULL);
 
 
 --
 -- Data for Name: schools; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO schools VALUES (2017, 1, 'Default School', '500 S. Street St.', 'Springfield', 'IL', '62704', NULL, 'Mr. Principal', 'www.rosariosis.org', NULL, NULL, 4, NULL);
+INSERT INTO schools VALUES (2018, 1, 'Default School', '500 S. Street St.', 'Springfield', 'IL', '62704', NULL, 'Mr. Principal', 'www.rosariosis.org', NULL, NULL, 4, NULL);
 
 
 --
 -- Data for Name: staff; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO staff VALUES (2017, 1, 1, NULL, 'Admin', 'Administrator', 'A', NULL, 'admin', '$6$dc51290a001671c6$97VSmw.Qu9sL6vpctFh62/YIbbR6b3DstJJxPXal2OndrtFszsxmVhdQaV2mJvb6Z38sPACXqDDQ7/uquwadd.', NULL, NULL, 'admin', NULL, ',1,', NULL, NULL, 1, NULL);
-INSERT INTO staff VALUES (2017, 2, 1, NULL, 'Teach', 'Teacher', 'T', NULL, 'teacher', '$6$cf0dc4c40d38891f$FqKT6nlTer3ujAf8CcQi6ABIEtlow0Va2p6HYh.M6eGWUfpgLr/pfrSwdIcTlV1LDxLg52puVETGMCYKL3vOo/', NULL, NULL, 'teacher', NULL, ',1,', NULL, NULL, 2, NULL);
-INSERT INTO staff VALUES (2017, 3, 1, NULL, 'Parent', 'Parent', 'P', NULL, 'parent', '$6$947c923597601364$Kgbb0Ey3lYTYnqM66VkFRgJVFDW48cBAfNF7t0CVjokL7drcEFId61whqpLrRI1w0q2J2VPfg86Obaf1tG2Ng1', NULL, NULL, 'parent', NULL, NULL, NULL, NULL, 3, NULL);
+INSERT INTO staff VALUES (2018, 1, 1, NULL, 'Admin', 'Administrator', 'A', NULL, 'admin', '$6$dc51290a001671c6$97VSmw.Qu9sL6vpctFh62/YIbbR6b3DstJJxPXal2OndrtFszsxmVhdQaV2mJvb6Z38sPACXqDDQ7/uquwadd.', NULL, NULL, 'admin', NULL, ',1,', NULL, NULL, 1, NULL);
+INSERT INTO staff VALUES (2018, 2, 1, NULL, 'Teach', 'Teacher', 'T', NULL, 'teacher', '$6$cf0dc4c40d38891f$FqKT6nlTer3ujAf8CcQi6ABIEtlow0Va2p6HYh.M6eGWUfpgLr/pfrSwdIcTlV1LDxLg52puVETGMCYKL3vOo/', NULL, NULL, 'teacher', NULL, ',1,', NULL, NULL, 2, NULL);
+INSERT INTO staff VALUES (2018, 3, 1, NULL, 'Parent', 'Parent', 'P', NULL, 'parent', '$6$947c923597601364$Kgbb0Ey3lYTYnqM66VkFRgJVFDW48cBAfNF7t0CVjokL7drcEFId61whqpLrRI1w0q2J2VPfg86Obaf1tG2Ng1', NULL, NULL, 'parent', NULL, NULL, NULL, NULL, 3, NULL);
 
 
 --
@@ -4343,19 +4343,19 @@ INSERT INTO staff_field_categories VALUES (3, 'Food Service', 3, NULL, 'Food_Ser
 -- Data for Name: student_enrollment; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO student_enrollment VALUES (1, 2017, 1, 1, 7, '2017-06-12', NULL, 3, NULL, 1, 1, 1);
+INSERT INTO student_enrollment VALUES (1, 2018, 1, 1, 7, '2018-06-18', NULL, 3, NULL, 1, 1, 1);
 
 
 --
 -- Data for Name: student_enrollment_codes; Type: TABLE DATA; Schema: public; Owner: rosariosis
 --
 
-INSERT INTO student_enrollment_codes VALUES (1, 2017, 'Moved from District', 'MOVE', 'Drop', NULL, 1);
-INSERT INTO student_enrollment_codes VALUES (2, 2017, 'Expelled', 'EXP', 'Drop', NULL, 2);
-INSERT INTO student_enrollment_codes VALUES (3, 2017, 'Beginning of Year', 'EBY', 'Add', 'Y', 3);
-INSERT INTO student_enrollment_codes VALUES (4, 2017, 'From Other District', 'OTHER', 'Add', NULL, 4);
-INSERT INTO student_enrollment_codes VALUES (5, 2017, 'Transferred in District', 'TRAN', 'Drop', NULL, 5);
-INSERT INTO student_enrollment_codes VALUES (6, 2017, 'Transferred in District', 'EMY', 'Add', NULL, 6);
+INSERT INTO student_enrollment_codes VALUES (1, 2018, 'Moved from District', 'MOVE', 'Drop', NULL, 1);
+INSERT INTO student_enrollment_codes VALUES (2, 2018, 'Expelled', 'EXP', 'Drop', NULL, 2);
+INSERT INTO student_enrollment_codes VALUES (3, 2018, 'Beginning of Year', 'EBY', 'Add', 'Y', 3);
+INSERT INTO student_enrollment_codes VALUES (4, 2018, 'From Other District', 'OTHER', 'Add', NULL, 4);
+INSERT INTO student_enrollment_codes VALUES (5, 2018, 'Transferred in District', 'TRAN', 'Drop', NULL, 5);
+INSERT INTO student_enrollment_codes VALUES (6, 2018, 'Transferred in District', 'EMY', 'Add', NULL, 6);
 
 
 --