+++ CHANGES.md
@@ -22,6 +22,7 @@ Changes in 3.9
 - Fix SQL error no AMOUNT when Print after Save Payments in StudentPayments.php & StaffPayments.php
 - Translate Help_en.php help texts in locale/[code].utf8/LC_MESSAGES/help.po & help.mo
 - Fix #244: Add DISPLAY_NAME to CONFIG table for every school in Configuration.php, rosariosis.sql & Update.fnc.php, sponsored by Asian Hope
+- Default school year is 2018 in rosariosis.sql & config.inc.sample.php
 
 Changes in 3.8
 --------------